fix: return null scores for non-applicable scorers
Multi-turn evals now return score: null for single-turn scorers
(Communication, Tone, etc.) so Braintrust shows N/A instead of 0%.
Single-turn evals return score: null for AssertionsPassed.

Extracted shared multi-turn scoring into scores.ts to deduplicate
across all 4 eval entry points.
